+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Change log
 - Increased the test coverage.
 - Added checks for the presence and correctness of the ``SITES`` setting.
 - Added Python 3.10, Django 4.0rc1 to the CI.
+- Changed the code to raise a ``DisallowedHost`` exception instead of a 404 if
+  unable to find a matching host for the current request.
 
 
 `0.0.1`_ (2021-09-14)